разбиение на модули, CommonJS
